refactor: rename menu files by Menu
